liburing.h: fix 32-bit compile warning
Cast 'ts' to uintptr_t, to avoid having the compiler complain about
casting it to a 64-bit size.

../src/include/liburing.h: In function ‘io_uring_prep_timeout_update’:
../src/include/liburing.h:342:42: warning: cast from pointer to integer of different size [-Wpointer-to-int-cast]
  342 |     (void *)(unsigned long)user_data, 0, (__u64)ts);
      |                                          ^

+++ src/include/liburing.h
@@ -339,7 +339,8 @@ static inline void io_uring_prep_timeout_update(struct io_uring_sqe *sqe,
 						__u64 user_data, unsigned flags)
 {
 	io_uring_prep_rw(IORING_OP_TIMEOUT_REMOVE, sqe, -1,
-				(void *)(unsigned long)user_data, 0, (__u64)ts);
+				(void *)(unsigned long)user_data, 0,
+				(uintptr_t)ts);
 	sqe->timeout_flags = flags | IORING_TIMEOUT_UPDATE;
 }
